Move the RemoteConsole its subclasses under ClientAdapter
