+++ lib/remote_console/client_adapter/regular_socket.rb
@@ -1,0 +1,13 @@
+module RemoteConsole
+  module ClientAdapter
+    class RegularSocket < Base
+      def fetch(length)
+        yield(@sock.recv_nonblock(length))
+      end
+
+      def issue(data)
+        @sock.write(data)
+      end
+    end
+  end
+end
