--- lib/remote_console/web_mks_legacy.rb
@@ -1,17 +1,0 @@
-module RemoteConsole
-  class WebMKSLegacy < WebMKS
-    def issue(data)
-      @driver.frame(data)
-    end
-
-    private
-
-    def protocol
-      'uint8utf8'.freeze
-    end
-
-    def path
-      @record.url
-    end
-  end
-end
