+++ spec/lib/remote_console/client_adapter/web_mks_legacy_spec.rb
@@ -1,4 +1,4 @@
-describe RemoteConsole::WebMKSLegacy do
+describe RemoteConsole::ClientAdapter::WebMKSLegacy do
   let(:record) { FactoryBot.create(:system_console, :url => '/12345') }
   subject { described_class.new(record, nil) }
