+++ spec/lib/remote_console/client_adapter/web_mks_spec.rb
@@ -1,4 +1,4 @@
-describe RemoteConsole::WebMKS do
+describe RemoteConsole::ClientAdapter::WebMKS do
   let(:record) { FactoryBot.create(:system_console, :secret => '12345') }
   subject { described_class.new(record, nil) }
